fix: Mudando de double/float pra decimal, para melhor usabilidade a longo prazo.

+++ Models/RelatorioGeral.cs
@@ -1,7 +1,7 @@
 namespace DesafioTecnicoC.Models;
 public record RelatorioGeral
 {
-  public float Total_Geral_Receitas {get;}
-  public float Total_Geral_Despesas {get;}
-  public float Saldo_Liquido_Geral {get;}
+  public decimal Total_Geral_Receitas {get;}
+  public decimal Total_Geral_Despesas {get;}
+  public decimal Saldo_Liquido_Geral {get;}
 }
