+++ Models/Transacao.cs
@@ -4,7 +4,7 @@ public record Transacao
 {
     public int Id {get;init;}
     public string Descricao { get; init; } = string.Empty;
-    public double Valor { get; init; }
+    public decimal Valor { get; init; }
     public TipoTransacao Tipo { get; init; }
     public int Id_Pessoa { get; init; }
 }
